alma 1/requirements.txt
# Core Data Analysis
pandas==2.0.3
numpy==1.24.3

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
wordcloud==1.8.2.2

# NLP & Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.4.1
spacy==3.5.3
python-dotenv==1.0.0

# Database Connections
psycopg2-binary==2.9.6
neo4j==5.8.1
py2neo==2021.2.3

# API & Web Scraping
tweepy==4.12.1
praw==7.7.0
requests==2.28.2

# Utilities
tqdm==4.65.0
python-dateutil==2.8.2